Propagate annotations from Service within workload cluster to Service within Supervisor namespace
The behavior depends on a new vSphere CPI configuration variable svcAnnoPropEnabled, whose default value is false. When the variable is set to true, vSphere CPI will propagate the annotations on Services to the VMServices in the Supervisor cluster. Otherwise, it will maintain the current behavior and not propagate the Service annotations.